Don't bundle gtk3 in appimages.
In practice this just makes the appimage crash when calling gtk3
stuff.  And since upgrading rfd, we no longer need it on most Linux
distros.  GTK3 is still used as a fallback, however, in which case
the app image will end up using the system GTK3 libs, which have a
chance of working.