Tonemapper: use a single wide color space for saturation computation.
This should hopefully give consistent results using the same parameters
in different target gamuts.  Although this is untested.

This also splits the desaturation power parameter into two parameters:
one for originally (in the source colors) desaturated colors, and one
for highly saturated colors.  This gives a lot more control to dial in
a good look.