Implement a poor-man's gamut clipper.
This can also serve as the basis for a poor-man's gamut compressor
in the future.

The reason for this is that OCIO's built-in gamut compressor is
just not good, to the extent that a hacked together poor-man's
gamut clipper/compressor is actually *better*.  The built-in one
shifts out-of-gamut colors towards the primaries, and just generally
isn't much better than a per-channel clip aside from being softer.
Really disappointing.  So, we're going the poor-man's route until
OCIO gets an even half-way decent gamut compressor.