Simplify the parameters for the tonemap curve.
Just one parameter each for the toe and shoulder.  This makes it a lot
more intuitive and less fiddly to tweak.  It removes a little bit of
flexibility, but in testing this can still achieve a wide range of
looks, and this isn't meant for final color grading anyway.

This also slightly changes the tonemap curve itself, now using a small
linear segment at the start of the shoulder curve.  This is a simpler
and more reliable way of keeping the shoulder from unduly affecting the
mids and darks, and in practice the second-order discontinuity doesn't
seem to have negative visual effects (e.g. gradient crawling).